requirements.txt
flask~=2.3.2
watchdog~=4.0.1
docx~=0.2.4
pillow~=10.4.0
pdfplumber
openai~=0.28.1
tesseract
pypdf
llama-index
pymongo~=4.8.0
Werkzeug~=3.0.1
Authlib~=1.3.1
pytesseract~=0.3.10
fitz~=0.0.1.dev2
langchain_community